+++ packages/paste-core/components/popover/stories/index.stories.tsx
@@ -1,7 +1,9 @@
 import * as React from 'react';
 import {Box} from '@twilio-paste/box';
+import {Button} from '@twilio-paste/button';
+import {Stack} from '@twilio-paste/stack';
 import {Text} from '@twilio-paste/text';
-import {Popover, PopoverContainer, PopoverButton} from '../src';
+import {usePopoverState, Popover, PopoverContainer, PopoverButton} from '../src';
 
 // eslint-disable-next-line import/no-default-export
 export default {
@@ -14,7 +16,7 @@ export default {
   },
 };
 
-export const Default = (): React.ReactNode => {
+export const Default: React.FC = () => {
   return (
     <Box height="300px">
       <PopoverContainer baseId="test-id" visible>
@@ -27,12 +29,12 @@ export const Default = (): React.ReactNode => {
   );
 };
 
-export const PopoverTop = (): React.ReactNode => {
+export const PopoverTop: React.FC = () => {
   return (
     <Box bottom={12} position="absolute">
       <PopoverContainer baseId="test-id" placement="top-start">
         <PopoverButton variant="primary">Open popover</PopoverButton>
-        <Popover aria-label="Popover">
+        <Popover aria-label="Popover" data-testid="top-popover">
           <Text as="span">This is the Twilio styled popover that you can use in all your applications.</Text>
         </Popover>
       </PopoverContainer>
@@ -40,7 +42,7 @@ export const PopoverTop = (): React.ReactNode => {
   );
 };
 
-export const PopoverLeft = (): React.ReactNode => {
+export const PopoverLeft: React.FC = () => {
   return (
     <Box position="absolute" right={12} top={12}>
       <PopoverContainer baseId="test-id" placement="left-start">
@@ -53,7 +55,7 @@ export const PopoverLeft = (): React.ReactNode => {
   );
 };
 
-export const PopoverRight = (): React.ReactNode => {
+export const PopoverRight: React.FC = () => {
   return (
     <PopoverContainer baseId="test-id" placement="right-start">
       <PopoverButton variant="primary">Open popover</PopoverButton>
@@ -64,7 +66,7 @@ export const PopoverRight = (): React.ReactNode => {
   );
 };
 
-export const PopoverBottom = (): React.ReactNode => {
+export const PopoverBottom: React.FC = () => {
   return (
     <PopoverContainer baseId="test-id" placement="bottom-start">
       <PopoverButton variant="primary">Open popover</PopoverButton>
@@ -72,5 +74,27 @@ export const PopoverBottom = (): React.ReactNode => {
         <Text as="span">This is the Twilio styled popover that you can use in all your applications.</Text>
       </Popover>
     </PopoverContainer>
+  );
+};
+
+export const StateHookExample: React.FC = () => {
+  const popover = usePopoverState({baseId: 'test-id'});
+  return (
+    <Box height="300px">
+      <Stack orientation="horizontal" spacing="space60">
+        <PopoverContainer state={popover}>
+          <PopoverButton variant="primary">Open popover</PopoverButton>
+          <Popover aria-label="Popover" data-testid="state-hook-popover">
+            <Text as="span">This is the Twilio styled popover that you can use in all your applications.</Text>
+          </Popover>
+        </PopoverContainer>
+        <Button variant="primary" onClick={() => popover.show()} data-testid="show-button">
+          Open popover on click
+        </Button>
+        <Button variant="primary" onClick={() => popover.hide()} data-testid="hide-button">
+          Close popover on click
+        </Button>
+      </Stack>
+    </Box>
   );
 };