+++ packages/paste-core/components/tabs/src/TabPanel.tsx
@@ -1,7 +1,17 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
+import {Box} from '@twilio-paste/box';
+import type {BoxStyleProps} from '@twilio-paste/box';
 import {TabPrimitivePanel} from '@twilio-paste/tabs-primitive';
 import {TabsContext} from './TabsContext';
+
+export const tabPanelStyles = {
+  borderRadius: 'borderRadius20',
+  _focusVisible: {
+    boxShadow: 'shadowFocus',
+    outline: 'none',
+  },
+} as Partial<BoxStyleProps>;
 
 export interface TabPanelProps {
   id?: string | undefined;
@@ -12,7 +22,7 @@ export interface TabPanelProps {
 const TabPanel = React.forwardRef<HTMLDivElement, TabPanelProps>(({children, ...props}, ref) => {
   const tab = React.useContext(TabsContext);
   return (
-    <TabPrimitivePanel {...(tab as any)} {...props} ref={ref}>
+    <TabPrimitivePanel {...(tab as any)} {...tabPanelStyles} {...props} as={Box} ref={ref}>
       {children}
     </TabPrimitivePanel>
   );